kaggle-scrapper/requirements.txt
selenium==4.21.0
webdriver-manager==4.0.1
requests==2.31.0
psycopg2-binary==2.9.9
